MACOS 10.3.9 port
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@388 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/layout/SpringElectrical.cpp
@@ -121,7 +121,7 @@ bool SpringElectrical::run() {
     layoutProxy->setNodeValue(n, inputLayout->getNodeValue(n));
     sizeNorm.set(n.id, inputSize->getNodeValue(n).norm()/2.0);
   }
-  double maxforce   = sqrt(graph->numberOfNodes());
+  double maxforce   = sqrt((float) graph->numberOfNodes());
   double temperature = 1.0;
   bool overlapTest=false;
